1.增大清除按钮的点击范围； 2.完善地址簿功能，解决了一些已知的bug； 3.修改所有交易记录页面的点击事件

+++ app/src/main/java/com/hualianzb/biut/utils/UiHelper.java
@@ -204,7 +204,7 @@ public class UiHelper {
         activity.startActivityForResult(intent, 101);
     }
 
-    public static void startChangeAddressBookActys(Context activity, int bookIndex) {
+    public static void startChangeAddressBookActys(Context activity, long bookIndex) {
         Intent intent = new Intent(activity, ChangeAddressActivity.class);
         intent.putExtra("bookIndex", bookIndex);
         activity.startActivity(intent);
